# contracts/002-trader-orders-views/openapi.yaml
openapi: 3.0.3
info:
  title: Money Market Order Processing API
  description: |
    Contract-first HTTP API for Money Market order processing. Baseline intake and lifecycle
    operations from `001-mm-order-processing`; **002-trader-orders-views** adds workspace-scoped
    list surfaces (`/term/*`, `/oncall/*`) so Term and OnCall orders are never mixed on the same
    list response. Workspace **Assigned** lists are **desk-wide** (all assignees) per User Story 2.
    Human-readable mirror: [api-v1.md](./api-v1.md). Trader-facing operations require header
    `X-User-Id` (actor identity; Assigned workspace lists are not filtered by assignee).
  version: 2.0.0
servers:
  - url: /
    description: Relative to deployment base (e.g. http://localhost:8080)
...
paths:
...
  /api/v1/orders/term/received:
    get:
      tags: [Orders]
      summary: List received Term orders
      description: |
        By default (`receivedView=NEAR_TERM`) returns only orders whose `valueDate` falls within **today and the next two calendar days** inclusive in the business calendar (**Europe/Paris**). Use `receivedView=ALL` for the full RECEIVED cohort for this workspace type (session UX may pass this from a trader “show all” control).
      operationId: listReceivedTermOrders
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - $ref: "#/components/parameters/ReceivedListView"
      responses:
        "200":
          description: Paged Term orders in RECEIVED status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderSummaryPage"
  /api/v1/orders/oncall/received:
    get:
      tags: [Orders]
      summary: List received OnCall orders
      description: |
        Same near-term window behaviour as Term Received; see `GET /api/v1/orders/term/received`.
      operationId: listReceivedOnCallOrders
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - $ref: "#/components/parameters/ReceivedListView"
      responses:
        "200":
          description: Paged OnCall orders in RECEIVED status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderSummaryPage"
  /api/v1/orders/term/assigned:
    get:
      tags: [Orders]
      summary: List all assigned Term orders (desk-wide)
      description: |
        Returns every order in **ASSIGNED** status with `orderType` **TERM** (all assignees).
        `X-User-Id` identifies the caller and is required for consistency with other Trader
        operations; the response is **not** restricted to orders assigned to that trader.
        Execute and unassign remain assignee-gated on their respective operations.
      operationId: listAssignedTermOrders
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Paged Term orders in ASSIGNED status (desk-wide)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderSummaryPage"
  /api/v1/orders/oncall/assigned:
    get:
      tags: [Orders]
      summary: List all assigned OnCall orders (desk-wide)
      description: |
        Returns every order in **ASSIGNED** status with `orderType` **ON_CALL** (all assignees).
        `X-User-Id` identifies the caller and is required for consistency; the response is **not**
        restricted to orders assigned to that trader.
      operationId: listAssignedOnCallOrders
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Paged OnCall orders in ASSIGNED status (desk-wide)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderSummaryPage"
  /api/v1/orders/term/executed:
    get:
      tags: [Orders]
      summary: List executed Term orders (workspace Executed view)
      operationId: listExecutedTermOrders
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Paged Term orders in EXECUTED status (US1 shell; accounting filters in later stories)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderSummaryPage"
  /api/v1/orders/oncall/executed:
    get:
      tags: [Orders]
      summary: List executed OnCall orders (workspace Executed view)
      operationId: listExecutedOnCallOrders
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Paged OnCall orders in EXECUTED status (US1 shell; accounting filters in later stories)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderSummaryPage"
...
  /api/v1/orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get order details
      operationId: getOrderDetails
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/OrderIdPath"
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetailsResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags: [Orders]
      summary: Update assigned order
      operationId: updateAssignedOrder
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/OrderIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetailsResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Caller is not the assigned Trader
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Invalid status for update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/orders/{orderId}/assign:
    post:
      tags: [Orders]
      summary: Assign order to requesting Trader
      operationId: assignOrder
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/OrderIdPath"
      responses:
        "200":
          description: Assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetailsResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Order not in RECEIVED status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/orders/{orderId}/unassign:
    post:
      tags: [Orders]
      summary: Unassign order
      operationId: unassignOrder
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/OrderIdPath"
      responses:
        "200":
          description: Unassigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetailsResponse"
        "403":
          description: Caller is not the assigned Trader
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Order not in ASSIGNED status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/orders/assigned:
    get:
      deprecated: true
      tags: [Orders]
      summary: List orders assigned to the requesting Trader
      description: |
        Deprecated — mixes Term and OnCall. Use `GET /api/v1/orders/term/assigned` or
        `GET /api/v1/orders/oncall/assigned` (002-trader-orders-views).
      operationId: listAssignedOrders
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Paged assigned orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderSummaryPage"
  /api/v1/orders/{orderId}/execute:
    post:
      tags: [Orders]
      summary: Record order execution
      operationId: executeOrder
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/OrderIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteOrderRequest"
      responses:
        "200":
          description: Executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetailsResponse"
        "400":
          description: Missing or invalid execution data, or executedRate below MinimumRate when order has a PM floor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Caller is not the assigned Trader
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Order not in ASSIGNED status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/orders/{orderId}/cancel:
    post:
      tags: [Orders]
      summary: Cancel received order
      operationId: cancelOrder
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/OrderIdPath"
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetailsResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Order not in RECEIVED status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/orders/{orderId}/reject:
    post:
      tags: [Orders]
      summary: Reject received or assigned order
      operationId: rejectOrder
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/OrderIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectOrderRequest"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetailsResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Caller is not the assigned Trader (when order is ASSIGNED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Order must be in RECEIVED or ASSIGNED status to reject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    UserIdHeader:
      name: X-User-Id
      in: header
      required: true
      description: MMXUser identity for Trader-facing and settings endpoints
      schema:
        type: string
        minLength: 1
    OrderIdPath:
      name: orderId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      description: Zero-based page index
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      description: Page size (max 100)
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 20
    ReceivedListView:
      name: receivedView
      in: query
      required: false
      description: |
        `NEAR_TERM` (default): only orders with `valueDate` in the inclusive three-calendar-day window (**today** through **today+2**) in **Europe/Paris**.
        `ALL`: every RECEIVED order for this workspace type (matches legacy unpaged-all behaviour).
      schema:
        $ref: "#/components/schemas/ReceivedListView"
  schemas:
    ErrorCode:
      type: string
      description: Machine-readable error code
      enum:
        - VALIDATION_ERROR
        - ORDER_NOT_FOUND
        - INVALID_STATUS_TRANSITION
        - UNAUTHORIZED_TRADER
        - DUPLICATE_ORDER
        - INSTITUTION_NOT_FOUND
        - ONCALL_SEGMENT_NOT_FOUND
        - ONCALL_PENDING_EXISTS
        - ONCALL_BACKDATED_VALUE_DATE
        - ONCALL_INVALID_SEGMENT_STATUS
        - ONCALL_SEGMENT_CANCELED
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
    ReceivedListView:
      type: string
      description: Received list horizon for workspace Received queues (US3)
      enum: [NEAR_TERM, ALL]
      default: NEAR_TERM
    OrderType:
      type: string
      enum: [TERM, ON_CALL]
    OrderOperation:
      type: string
      enum: [SUBSCRIPTION, INCREASE, DECREASE, REDEMPTION]
    OrderStatus:
      type: string
      enum: [RECEIVED, ASSIGNED, EXECUTED, ACCOUNTED, CANCELLED, REJECTED]
...
    OrderDetailsResponse:
      type: object
      required:
        - orderId
        - externalOrderReference
        - orderType
        - orderOperation
        - portfolioNumber
        - currency
        - amount
        - valueDate
        - status
        - createdAt
        - updatedAt
      properties:
        orderId:
          type: string
          format: uuid
        externalOrderReference:
          type: string
        orderType:
          $ref: "#/components/schemas/OrderType"
        orderOperation:
          $ref: "#/components/schemas/OrderOperation"
        portfolioNumber:
          type: string
        currency:
          type: string
        amount:
          type: number
          format: double
        valueDate:
          type: string
          format: date
        minimumRate:
          type: number
          format: double
          nullable: true
          description: Optional PM execution floor when supplied at intake; immutable after reception.
        tenor:
          type: string
          nullable: true
        noticePeriod:
          type: string
          nullable: true
        sourceContractNumber:
          type: string
          nullable: true
          description: >
            Intake contract reference for lifecycle operations (INCREASE, DECREASE, REDEMPTION); null for Subscription
            (PM-supplied value is discarded at reception). After execute on lifecycle orders, echoes the same reference
            as used for execution contract identity.
        status:
          $ref: "#/components/schemas/OrderStatus"
        assignedTraderId:
          type: string
          nullable: true
        assignedAt:
          type: string
          format: date-time
          nullable: true
        executedRate:
          type: number
          format: double
          nullable: true
        counterparty:
          type: string
          nullable: true
          description: Institution display name derived from intake institutionCode.
        institutionCode:
          type: string
          nullable: true
          maxLength: 32
          description: Catalog institution code selected by Portfolio Management at intake.
        executionTime:
          type: string
          format: date-time
          nullable: true
        dealingReference:
          type: string
          nullable: true
        generatedContractNumber:
          type: string
          nullable: true
          description: >
            After execution: new Money Market contract reference allocated by the system for SUBSCRIPTION orders;
            for INCREASE, DECREASE, and REDEMPTION equals the persisted intake sourceContractNumber (same underlying contract).
        rejectionReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderSummaryPage:
      type: object
      required: [content, totalElements, page, size]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/OrderSummaryResponse"
        totalElements:
          type: integer
          format: int64
          minimum: 0
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
    UpdateOrderRequest:
      type: object
      description: Partial update — only Amount and ValueDate may be changed. MinimumRate, institutionCode, and counterparty are intake-only and not mutable via this endpoint.
      properties:
        amount:
          type: number
          format: double
          exclusiveMinimum: 0
        valueDate:
          type: string
          format: date
    ExecuteOrderRequest:
      type: object
      required: [executedRate]
      properties:
        executedRate:
          type: number
          format: double
          minimum: 0
...
    RejectOrderRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 500